+++ scann/scann/base/single_machine_base.h
@@ -269,13 +269,13 @@ class SingleMachineSearcherBase : public UntypedSingleMachineSearcherBase {
   virtual Status FindNeighbors(const DatapointPtr<T>& query,
                                const SearchParameters& params,
                                NNResultsVector* result,
-                               float* SOW = nullptr,
+                               unsigned long long int* SOW = nullptr,
                                size_t begin = 0,
                                size_t curSize = 0) const;
 
   Status FindNeighbors(const DatapointPtr<T>& query,
                        NNResultsVector* result,
-                       float* SOW = nullptr,
+                       unsigned long long int* SOW = nullptr,
                        size_t begin = 0,
                        size_t curSize = 0) const {
     return FindNeighbors(query, default_search_parameters_, result, SOW, begin, curSize);
@@ -284,25 +284,25 @@ class SingleMachineSearcherBase : public UntypedSingleMachineSearcherBase {
   Status FindNeighborsNoSortNoExactReorder(const DatapointPtr<T>& query,
                                            const SearchParameters& params,
                                            NNResultsVector* result,
-                                           float* SOW = nullptr,
+                                           unsigned long long int* SOW = nullptr,
                                            size_t begin = 0,
                                            size_t curSize = 0) const;
 
   Status FindNeighborsBatched(const TypedDataset<T>& queries,
                               MutableSpan<NNResultsVector> results,
-                              float* SOW = nullptr,
+                              unsigned long long int* SOW = nullptr,
                               size_t begin = 0,
                               size_t curSize = 0) const;
   Status FindNeighborsBatched(const TypedDataset<T>& queries,
                               ConstSpan<SearchParameters> params,
                               MutableSpan<NNResultsVector> results,
-                              float* SOW = nullptr,
+                              unsigned long long int* SOW = nullptr,
                               size_t begin = 0,
                               size_t curSize = 0) const;
   Status FindNeighborsBatchedNoSortNoExactReorder(
       const TypedDataset<T>& queries, ConstSpan<SearchParameters> params,
       MutableSpan<NNResultsVector> results,
-      float* SOW = nullptr,
+      unsigned long long int* SOW = nullptr,
       size_t begin = 0,
       size_t curSize = 0) const;
 
@@ -494,14 +494,14 @@ class SingleMachineSearcherBase : public UntypedSingleMachineSearcherBase {
   virtual Status FindNeighborsImpl(const DatapointPtr<T>& query,
                                    const SearchParameters& params,
                                    NNResultsVector* result,
-                                   float* SOW = nullptr,
+                                   unsigned long long int* SOW = nullptr,
                                    size_t begin = 0,
                                    size_t curSize = 0) const = 0;
 
   virtual Status FindNeighborsBatchedImpl(
       const TypedDataset<T>& queries, ConstSpan<SearchParameters> params,
       MutableSpan<NNResultsVector> results,
-      float* SOW = nullptr,
+      unsigned long long int* SOW = nullptr,
       size_t begin = 0,
       size_t curSize = 0) const;
