+++ scann/scann/scann_ops/cc/scann.h
@@ -56,20 +56,20 @@ class ScannInterface {
 
   Status Search(const DatapointPtr<float> query, NNResultsVector* res,
                 int final_nn, int pre_reorder_nn, int leaves,
-                float* SOW = nullptr,
+                unsigned long long int* SOW = nullptr,
                 size_t begin = 0,
                 size_t curSize = 0) const; // [ANNA] SOW
   Status SearchBatched(const DenseDataset<float>& queries,
                        MutableSpan<NNResultsVector> res, int final_nn,
                        int pre_reorder_nn, int leaves,
-                       float* SOW = nullptr,
+                       unsigned long long int* SOW = nullptr,
                        size_t begin = 0,
                        size_t curSize = 0) const;
   Status SearchBatchedParallel(const DenseDataset<float>& queries,
                                MutableSpan<NNResultsVector> res, int final_nn,
                                int pre_reorder_nn, int leaves,
                                int batch_size,
-                               float* SOW = nullptr,
+                               unsigned long long int* SOW = nullptr,
                                size_t begin = 0,
                                size_t curSize = 0) const;   // [ANNA] batch_size, SOW
   Status Serialize(std::string path, std::string coarse_path, bool load_coarse, std::string fine_path, bool load_fine);
